+++ core/src/androidTest/java/io/snabble/sdk/EncodedCodesGeneratorTest.java
@@ -360,6 +360,34 @@ public class EncodedCodesGeneratorTest extends SnabbleSdkTest {
         Assert.assertEquals("snabble;\n1;asdf123\n1000;8715700421698\n7;4008400301020", codes.get(0));
     }
 
+    @Test
+    @UiThreadTest
+    public void testCSVv2Format() {
+        EncodedCodesOptions options = new EncodedCodesOptions.Builder(project)
+                .prefix("snabble;{qrCodeIndex};{qrCodeCount}\n")
+                .separator("\n")
+                .suffix("")
+                .repeatCodes(false)
+                .countSeparator(";")
+                .maxCodes(2)
+                .build();
+
+        EncodedCodesGenerator generator = new EncodedCodesGenerator(options);
+
+        Product duplo = project.getProductDatabase().findBySku("49");
+        addToCart(duplo, 7, ScannedCode.parseDefault(project, "4008400301020"));
+
+        Product heinz = project.getProductDatabase().findBySku("42");
+        addToCart(heinz, 1000, ScannedCode.parseDefault(project, "8715700421698"));
+        generator.add("asdf123");
+        generator.add(project.getShoppingCart());
+
+        ArrayList<String> codes = generator.generate();
+        Assert.assertEquals(2, codes.size());
+        Assert.assertEquals("snabble;1;2\n1;asdf123\n1000;8715700421698", codes.get(0));
+        Assert.assertEquals("snabble;2;2\n7;4008400301020", codes.get(1));
+    }
+
     private void addToCart(Product product, int quantity, ScannedCode scannedCode) {
         ShoppingCart cart = project.getShoppingCart();
         ShoppingCart.Item item = cart.newItem(product, scannedCode);