+++ core/src/main/java/io/snabble/sdk/Project.java
@@ -52,6 +52,7 @@ public class Project {
     private Shop checkedInShop;
     private CustomerCardInfo[] acceptedCustomerCardInfos;
     private CustomerCardInfo requiredCustomerCardInfo;
+    private PaymentMethod[] availablePaymentMethods;
 
     private Map<String, String> urls;
 
@@ -67,6 +68,7 @@ public class Project {
 
     private int maxOnlinePaymentLimit;
     private int maxCheckoutLimit;
+    private JsonObject encodedCodesJsonObject;
 
     Project(JsonObject jsonObject) throws IllegalArgumentException {
         Snabble snabble = Snabble.getInstance();
@@ -129,21 +131,11 @@ public class Project {
 
         isCheckoutAvailable = JsonUtils.getBooleanOpt(jsonObject, "enableCheckout", true);
 
-        if (jsonObject.has("encodedCodes")) {
-            JsonElement encodedCodes = jsonObject.get("encodedCodes");
+        if (jsonObject.has("qrCodeOffline")) {
+            JsonElement encodedCodes = jsonObject.get("qrCodeOffline");
             if (!encodedCodes.isJsonNull()) {
-                JsonObject object = encodedCodes.getAsJsonObject();
-
-                encodedCodesOptions = new EncodedCodesOptions.Builder(this)
-                        .prefix(JsonUtils.getStringOpt(object, "prefix", ""))
-                        .separator(JsonUtils.getStringOpt(object, "separator", "\n"))
-                        .suffix(JsonUtils.getStringOpt(object, "suffix", ""))
-                        .maxCodes(JsonUtils.getIntOpt(object, "maxCodes", 100))
-                        .finalCode(JsonUtils.getStringOpt(object, "finalCode", ""))
-                        .nextCode(JsonUtils.getStringOpt(object, "nextCode", ""))
-                        .nextCodeWithCheck(JsonUtils.getStringOpt(object, "nextCodeWithCheck", ""))
-                        .maxSizeMm(JsonUtils.getIntOpt(object, "maxSizeMM", -1))
-                        .build();
+                encodedCodesJsonObject = encodedCodes.getAsJsonObject();
+                encodedCodesOptions = EncodedCodesOptions.fromJsonObject(this, encodedCodesJsonObject);
             }
         }
 
@@ -184,6 +176,16 @@ public class Project {
                 requiredCustomerCardInfo = customerCardInfo;
             }
         }
+
+        List<PaymentMethod> paymentMethodList = new ArrayList<>();
+        String[] paymentMethods = JsonUtils.getStringArrayOpt(jsonObject, "paymentMethods", new String[0]);
+        for (String paymentMethod : paymentMethods) {
+            PaymentMethod pm = PaymentMethod.fromString(paymentMethod);
+            if (pm != null) {
+                paymentMethodList.add(pm);
+            }
+        }
+        availablePaymentMethods = paymentMethodList.toArray(new PaymentMethod[paymentMethodList.size()]);
 
         if (jsonObject.has("shops")) {
             shops = Shop.fromJson(jsonObject.get("shops"));
@@ -434,10 +436,18 @@ public class Project {
      */
     public void setCustomerCardId(String loyaltyCardId) {
         this.loyaltyCardId = loyaltyCardId;
+
+        if (encodedCodesJsonObject != null) {
+            encodedCodesOptions = EncodedCodesOptions.fromJsonObject(this, encodedCodesJsonObject);
+        }
     }
 
     public String getCustomerCardId() {
         return loyaltyCardId;
+    }
+
+    public PaymentMethod[] getAvailablePaymentMethods() {
+        return availablePaymentMethods;
     }
 
     public CodeTemplate getDefaultCodeTemplate() {