+++ ui/src/main/java/io/snabble/sdk/ui/cart/ShoppingCartView.java
@@ -29,6 +29,7 @@ import com.squareup.picasso.Picasso;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 import androidx.appcompat.widget.AppCompatImageView;
 import androidx.recyclerview.widget.DefaultItemAnimator;
@@ -65,6 +66,8 @@ public class ShoppingCartView extends FrameLayout implements Checkout.OnCheckout
     private Button pay;
     private View coordinatorLayout;
     private ViewGroup emptyState;
+    private View restore;
+    private TextView scanProducts;
     private Snackbar snackbar;
     private DelayedProgressDialog progressDialog;
     private boolean hasAnyImages;
@@ -123,7 +126,7 @@ public class ShoppingCartView extends FrameLayout implements Checkout.OnCheckout
     private void inflateView(Context context, AttributeSet attrs) {
         inflate(getContext(), R.layout.snabble_view_shopping_cart, this);
         picasso = Picasso.with(getContext());
-        Project project = SnabbleUI.getProject();
+        final Project project = SnabbleUI.getProject();
 
         if (cart != null) {
             cart.removeListener(shoppingCartListener);
@@ -178,12 +181,21 @@ public class ShoppingCartView extends FrameLayout implements Checkout.OnCheckout
         pay.setOnClickListener(new OneShotClickListener() {
             @Override
             public void click() {
-                checkout.checkout();
-                Telemetry.event(Telemetry.Event.ClickCheckout);
+                if (cart.hasReachedMaxCheckoutLimit()) {
+                    Project project = SnabbleUI.getProject();
+                    String message = getResources().getString(R.string.Snabble_limitsAlert_checkoutNotAvailable,
+                            project.getPriceFormatter().format(project.getMaxCheckoutLimit()));
+                    snackbar = UIUtils.snackbar(coordinatorLayout, message, UIUtils.SNACKBAR_LENGTH_VERY_LONG);
+                    snackbar.show();
+                } else {
+                    checkout.checkout();
+                    Telemetry.event(Telemetry.Event.ClickCheckout);
+                }
             }
         });
 
-        View scanProducts = findViewById(R.id.scan_products);
+
+        scanProducts = findViewById(R.id.scan_products);
         scanProducts.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -191,6 +203,14 @@ public class ShoppingCartView extends FrameLayout implements Checkout.OnCheckout
                 if (callback != null) {
                     callback.showScannerWithCode(null);
                 }
+            }
+        });
+
+        restore = findViewById(R.id.restore);
+        restore.setOnClickListener(new OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                cart.restore();
             }
         });
 
@@ -323,6 +343,14 @@ public class ShoppingCartView extends FrameLayout implements Checkout.OnCheckout
         } else {
             emptyState.setVisibility(View.VISIBLE);
             pay.setVisibility(View.GONE);
+        }
+
+        if (cart.isRestorable() && cart.getBackupTimestamp() > System.currentTimeMillis() - TimeUnit.MINUTES.toMillis(5)) {
+            restore.setVisibility(View.VISIBLE);
+            scanProducts.setText(R.string.Snabble_Shoppingcart_emptyState_restartButtonTitle);
+        } else {
+            restore.setVisibility(View.GONE);
+            scanProducts.setText(R.string.Snabble_Shoppingcart_emptyState_buttonTitle);
         }
     }